chore: improve FormControl exports
